% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolic.R
\docType{data}
\name{metabolic}
\alias{metabolic}
\title{metabolic syndrome dataset
Subjects of a metabolic syndrome study}
\format{
A dataframe with columns and rows
\describe{
 \item{ID}{ID}
 \item{RURAL}{Rural=1 or Urban=2}
 \item{DOB}{Date of Birth}
 \item{AGE}{Age of subjects}
 \item{DMDX}{History of Diabetes, 0=No, 1=Yes}
 \item{DISLIPID}{History of Dyslipidemia, 0=No, 1=Yes}
}
}
\usage{
metabolic
}
\description{
metabolic syndrome dataset
Subjects of a metabolic syndrome study
}
\keyword{datasets}
