% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroke.R
\docType{data}
\name{stroke}
\alias{stroke}
\title{Stroke dataset}
\format{
A dataframe with columns and rows
\describe{
 \item{sex}{sex of patient}
 \item{race}{race of patient}
 \item{doa}{date of admission}
 \item{doa}{date of discharge}
}
}
\usage{
stroke
}
\description{
Data from in-patients with acute stroke
}
\keyword{datasets}
