% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pup.R
\docType{data}
\name{pup}
\alias{pup}
\title{PUP dataset}
\format{
A dataframe with columns and rows
\describe{
 \item{no}{ID}
 \item{age}{Age for patient}
 \item{gender}{Gender for patient}
}
}
\usage{
pup
}
\description{
Data from Peptic Ulcer Disease patients
}
\keyword{datasets}
